Simplify API
Remove `topic?` and `partition?` from the core namespace as their usage
was too confusing. If dynamic creation of topics is enabled, those
functions would always retur true anyway. If disabled, the consumer
would indeed return false but the producer would hang for a very long
time and throw an exception.