Remove ?nil convention and improve consumer options
Nilable arguments and map keys were prefixed with '?'. The goal was to
make obvious to the user what could be nil. Although useful, it is not a
widely used convention and hence could actually harm the user.

Allowing the :listen option in the consumer options is bad practise.
There is a function for that and the user should use that instead.