Reorganize the API and prepare for 1.0.0
Milena has been renamed to dvlopt.kafka for clarity. The API regarding
Kafka itself has been improved in order to be more meaningful to use.
After a lot of trial and error, this organization has been chosen for
the first major release and is supposed to stick around.

The API regarding Kafka Streams has greatly matured and is now
documented almost everywhere as needed.

A few things remain to be finalized, some testing needs to be done, and
then everything will be ready for a stable release.